requirements.txt
appdirs==1.4.3
autopep8==1.5.3
beautifulsoup4==4.9.1
bs4==0.0.1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
MouseInfo==0.1.3
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
progress==1.5
pyaes==1.6.1
pyasn1==0.4.8
PyAutoGUI==0.9.50
pycodestyle==2.6.0
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pyparsing==2.4.6
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
pysqlite3==0.4.3
pytesseract==0.3.4
python3-xlib==0.15
pytoml==0.1.21
PyTweening==1.0.3
requests==2.22.0
retrying==1.3.3
rsa==4.6
selenium==3.141.0
six==1.14.0
soupsieve==2.0.1
Telethon==1.15.0
toml==0.10.1
urllib3==1.25.8
webencodings==0.5.1
